+++ packages/fiori/src/Wizard.ts
@@ -1,10 +1,10 @@
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import getEffectiveScrollbarStyle from "@ui5/webcomponents-base/dist/util/getEffectiveScrollbarStyle.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import ItemNavigation from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
@@ -12,7 +12,6 @@ import NavigationMode from "@ui5/webcomponents-base/dist/types/NavigationMode.js
 import clamp from "@ui5/webcomponents-base/dist/util/clamp.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
 import type { ResizeObserverCallback } from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
-import { isPhone } from "@ui5/webcomponents-base/dist/Device.js";
 import debounce from "@ui5/webcomponents-base/dist/util/debounce.js";
 import { getFirstFocusableElement } from "@ui5/webcomponents-base/dist/util/FocusableElements.js";
 import Button from "@ui5/webcomponents/dist/Button.js";
@@ -38,7 +37,7 @@ import WizardTab from "./WizardTab.js";
 import WizardStep from "./WizardStep.js";
 
 // Template and Styles
-import WizardTemplate from "./generated/templates/WizardTemplate.lit.js";
+import WizardTemplate from "./WizardTemplate.js";
 import WizardCss from "./generated/themes/Wizard.css.js";
 import WizardPopoverCss from "./generated/themes/WizardPopover.css.js";
 
@@ -92,7 +91,9 @@ type StepInfo = {
 	pos: number,
 	accInfo: AccessibilityInformation,
 	refStepId: string,
-	styles: object,
+	styles: {
+		zIndex: number
+	},
 }
 
 /**
@@ -184,7 +185,7 @@ type StepInfo = {
 	tag: "ui5-wizard",
 	languageAware: true,
 	fastNavigation: true,
-	renderer: litRender,
+	renderer: jsxRenderer,
 	styles: [
 		WizardCss,
 		WizardPopoverCss,
@@ -207,25 +208,14 @@ type StepInfo = {
  * @param {boolean} withScroll true when the event occurs due to user scrolling.
  * @public
  */
-@event<WizardStepChangeEventDetail>("step-change", {
-	detail: {
-		/**
-		* @public
-		*/
-		step: { type: HTMLElement },
-		/**
-		* @public
-		*/
-		previousStep: { type: HTMLElement },
-		/**
-		* @public
-		*/
-		withScroll: { type: Boolean },
-	},
+@event("step-change", {
 	bubbles: true,
 })
 
 class Wizard extends UI5Element {
+	eventDetails!: {
+		"step-change": WizardStepChangeEventDetail
+	}
 	/**
 	 * Defines how the content of the `ui5-wizard` would be visualized.
 	 * @public
@@ -337,19 +327,6 @@ class Wizard extends UI5Element {
 		});
 
 		this._onStepResize = this.onStepResize.bind(this);
-	}
-
-	get classes() {
-		return {
-			root: {
-				"ui5-wiz-root": true,
-			},
-			popover: {
-				"ui5-wizard-responsive-popover": true,
-				"ui5-wizard-popover": !isPhone(),
-				"ui5-wizard-dialog": isPhone(),
-			},
-		};
 	}
 
 	static get SCROLL_DEBOUNCE_RATE() {
@@ -469,7 +446,7 @@ class Wizard extends UI5Element {
 	 * **Note:** the handler is bound in the template.
 	 * @private
 	 */
-	onSelectionChangeRequested(e: MouseEvent) {
+	onSelectionChangeRequested(e: CustomEvent) {
 		this.selectionRequestedByClick = true;
 		this.changeSelectionByStepAction(e.target as WizardTab);
 	}
@@ -479,7 +456,7 @@ class Wizard extends UI5Element {
 	 * **Note:** the handler is bound in the template.
 	 * @private
 	 */
-	onScroll(e: MouseEvent) {
+	onScroll(e: Event) {
 		if (this.selectionRequestedByClick) {
 			this.selectionRequestedByClick = false;
 			return;
@@ -495,7 +472,7 @@ class Wizard extends UI5Element {
 	 * **Note:** the handler is bound in the template.
 	 * @private
 	 */
-	onStepInHeaderFocused(e: FocusEvent) {
+	onStepInHeaderFocused(e: CustomEvent) {
 		this._itemNavigation.setCurrentItem(e.target as WizardTab);
 	}
 
@@ -1037,7 +1014,7 @@ class Wizard extends UI5Element {
 				stepToSelect.selected = true;
 			}
 
-			this.fireDecoratorEvent<WizardStepChangeEventDetail>("step-change", {
+			this.fireDecoratorEvent("step-change", {
 				step: stepToSelect,
 				previousStep: selectedStep,
 				withScroll,