+++ packages/main/src/Button.ts
@@ -1,10 +1,10 @@
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import {
 	isSpace,
 	isEnter,
@@ -12,7 +12,7 @@ import {
 	isShift,
 } from "@ui5/webcomponents-base/dist/Keys.js";
 import { getEffectiveAriaLabelText } from "@ui5/webcomponents-base/dist/util/AccessibilityTextsHelper.js";
-import type { AccessibilityAttributes, PassiveEventListenerObject } from "@ui5/webcomponents-base/dist/types.js";
+import type { AccessibilityAttributes, AriaRole } from "@ui5/webcomponents-base";
 import type { ITabbable } from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import type { I18nText } from "@ui5/webcomponents-base/dist/i18nBundle.js";
@@ -25,12 +25,12 @@ import {
 import willShowContent from "@ui5/webcomponents-base/dist/util/willShowContent.js";
 import { submitForm, resetForm } from "@ui5/webcomponents-base/dist/features/InputElementsFormSupport.js";
 import { getEnableDefaultTooltips } from "@ui5/webcomponents-base/dist/config/Tooltips.js";
+import toLowercaseEnumValue from "@ui5/webcomponents-base/dist/util/toLowercaseEnumValue.js";
 import ButtonDesign from "./types/ButtonDesign.js";
 import ButtonType from "./types/ButtonType.js";
 import type ButtonAccessibleRole from "./types/ButtonAccessibleRole.js";
-import ButtonTemplate from "./generated/templates/ButtonTemplate.lit.js";
+import ButtonTemplate from "./ButtonTemplate.js";
 import Icon from "./Icon.js";
-import IconMode from "./types/IconMode.js";
 
 import { BUTTON_ARIA_TYPE_ACCEPT, BUTTON_ARIA_TYPE_REJECT, BUTTON_ARIA_TYPE_EMPHASIZED } from "./generated/i18n/i18n-defaults.js";
 
@@ -85,7 +85,7 @@ type ButtonAccessibilityAttributes = Pick<AccessibilityAttributes, "expanded" | 
 	tag: "ui5-button",
 	formAssociated: true,
 	languageAware: true,
-	renderer: litRender,
+	renderer: jsxRenderer,
 	template: ButtonTemplate,
 	styles: buttonCss,
 	dependencies: [Icon],
@@ -100,18 +100,22 @@ type ButtonAccessibilityAttributes = Pick<AccessibilityAttributes, "expanded" | 
  * @public
  * @native
  */
-@event("click", {
-	bubbles: true,
-})
+// @event("click", {
+// 	bubbles: true,
+// })
 /**
  * Fired whenever the active state of the component changes.
  * @private
  */
-@event("_active-state-change", {
+@event("active-state-change", {
 	bubbles: true,
 	cancelable: true,
 })
 class Button extends UI5Element implements IButton {
+	eventDetails!: {
+		"active-state-change": void
+		// click: void
+	}
 	/**
 	 * Defines the component design.
 	 * @default "Default"
@@ -327,8 +331,6 @@ class Button extends UI5Element implements IButton {
 
 	_deactivate: () => void;
 
-	_ontouchstart: PassiveEventListenerObject;
-
 	@i18n("@ui5/webcomponents")
 	static i18nBundle: I18nBundle;
 
@@ -345,19 +347,14 @@ class Button extends UI5Element implements IButton {
 
 			isGlobalHandlerAttached = true;
 		}
+	}
 
-		const handleTouchStartEvent = () => {
-			if (this.nonInteractive) {
-				return;
-			}
+	_ontouchstart() {
+		if (this.nonInteractive) {
+			return;
+		}
 
-			this._setActiveState(true);
-		};
-
-		this._ontouchstart = {
-			handleEvent: handleTouchStartEvent,
-			passive: true,
-		};
+		this._setActiveState(true);
 	}
 
 	onEnterDOM() {
@@ -449,7 +446,7 @@ class Button extends UI5Element implements IButton {
 	}
 
 	_setActiveState(active: boolean) {
-		const eventPrevented = !this.fireDecoratorEvent("_active-state-change");
+		const eventPrevented = !this.fireDecoratorEvent("active-state-change");
 
 		if (eventPrevented) {
 			return;
@@ -464,22 +461,6 @@ class Button extends UI5Element implements IButton {
 
 	get hasButtonType() {
 		return this.design !== ButtonDesign.Default && this.design !== ButtonDesign.Transparent;
-	}
-
-	get iconMode() {
-		if (!this.icon) {
-			return "";
-		}
-
-		return IconMode.Decorative;
-	}
-
-	get endIconMode() {
-		if (!this.endIcon) {
-			return "";
-		}
-
-		return IconMode.Decorative;
 	}
 
 	get isIconOnly() {
@@ -506,8 +487,8 @@ class Button extends UI5Element implements IButton {
 		return Button.i18nBundle.getText(Button.typeTextMappings()[this.design]);
 	}
 
-	get effectiveAccRole() {
-		return this.accessibleRole.toLowerCase();
+	get effectiveAccRole(): AriaRole {
+		return toLowercaseEnumValue(this.accessibleRole);
 	}
 
 	get tabIndexValue() {
@@ -518,10 +499,10 @@ class Button extends UI5Element implements IButton {
 		const tabindex = this.getAttribute("tabindex");
 
 		if (tabindex) {
-			return tabindex;
+			return Number.parseInt(tabindex);
 		}
 
-		return this.nonInteractive ? "-1" : this.forcedTabIndex;
+		return this.nonInteractive ? -1 : Number.parseInt(this.forcedTabIndex);
 	}
 
 	get showIconTooltip() {