+++ packages/main/src/TimeSelectionClocks.ts
@@ -1,6 +1,6 @@
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import "@ui5/webcomponents-localization/dist/features/calendar/Gregorian.js"; // default calendar for bundling
 import {
 	isDown,
@@ -28,7 +28,7 @@ import SegmentedButtonItem from "./SegmentedButtonItem.js";
 import type { TimePickerClockChangeEventDetail } from "./TimePickerClock.js";
 
 // Template
-import TimeSelectionClocksTemplate from "./generated/templates/TimeSelectionClocksTemplate.lit.js";
+import TimeSelectionClocksTemplate from "./TimeSelectionClocksTemplate.js";
 
 // Styles
 import TimeSelectionClocksCss from "./generated/themes/TimeSelectionClocks.css.js";
@@ -70,6 +70,10 @@ import TimeSelectionClocksCss from "./generated/themes/TimeSelectionClocks.css.j
 })
 
 class TimeSelectionClocks extends TimePickerInternals {
+	eventDetails!: TimePickerInternals["eventDetails"] & {
+		"close-picker": void,
+	};
+
 	/**
 	 * Flag for pressed Space key
 	 */