+++ packages/main/src/TimeSelectionInputs.ts
@@ -1,6 +1,6 @@
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
 import "@ui5/webcomponents-localization/dist/features/calendar/Gregorian.js"; // default calendar for bundling
 import {
 	isEnter,
@@ -20,7 +20,7 @@ import {
 } from "./generated/i18n/i18n-defaults.js";
 
 // Template
-import TimeSelectionInputsTemplate from "./generated/templates/TimeSelectionInputsTemplate.lit.js";
+import TimeSelectionInputsTemplate from "./TimeSelectionInputsTemplate.js";
 
 // Styles
 import TimeSelectionInputsCss from "./generated/themes/TimeSelectionInputs.css.js";
@@ -56,6 +56,10 @@ import TimeSelectionInputsCss from "./generated/themes/TimeSelectionInputs.css.j
 	bubbles: true,
 })
 class TimeSelectionInputs extends TimePickerInternals {
+	eventDetails!: TimePickerInternals["eventDetails"] & {
+		"close-inputs": void,
+	}
+
 	@property({ type: Number })
 	_editedInput = -1;