+++ packages/main/src/Tree.ts
@@ -3,11 +3,13 @@ import customElement from "@ui5/webcomponents-base/dist/decorators/customElement
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import DragRegistry from "@ui5/webcomponents-base/dist/util/dragAndDrop/DragRegistry.js";
+import handleDragOver from "@ui5/webcomponents-base/dist/util/dragAndDrop/handleDragOver.js";
+import handleDrop from "@ui5/webcomponents-base/dist/util/dragAndDrop/handleDrop.js";
 import { findClosestPosition } from "@ui5/webcomponents-base/dist/util/dragAndDrop/findClosestPosition.js";
 import Orientation from "@ui5/webcomponents-base/dist/types/Orientation.js";
 import MovePlacement from "@ui5/webcomponents-base/dist/types/MovePlacement.js";
-import event from "@ui5/webcomponents-base/dist/decorators/event.js";
-import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
+import event from "@ui5/webcomponents-base/dist/decorators/event-strict.js";
+import jsxRenderer from "@ui5/webcomponents-base/dist/renderer/JsxRenderer.js";
 import DropIndicator from "./DropIndicator.js";
 import TreeItem from "./TreeItem.js";
 import type TreeItemBase from "./TreeItemBase.js";
@@ -28,7 +30,7 @@ import type {
 } from "./List.js";
 
 // Template
-import TreeTemplate from "./generated/templates/TreeTemplate.lit.js";
+import TreeTemplate from "./TreeTemplate.js";
 
 // Styles
 import TreeCss from "./generated/themes/Tree.css.js";
@@ -104,7 +106,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  */
 @customElement({
 	tag: "ui5-tree",
-	renderer: litRender,
+	renderer: jsxRenderer,
 	styles: TreeCss,
 	template: TreeTemplate,
 	dependencies: [
@@ -123,13 +125,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @param {HTMLElement} item the toggled item.
  * @public
  */
-@event<TreeItemToggleEventDetail>("item-toggle", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("item-toggle", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -139,13 +135,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @since 1.0.0-rc.16
  * @public
  */
-@event<TreeItemMouseoverEventDetail>("item-mouseover", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("item-mouseover", {
 	bubbles: true,
 })
 /**
@@ -154,13 +144,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @since 1.0.0-rc.16
  * @public
  */
-@event<TreeItemMouseoutEventDetail>("item-mouseout", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("item-mouseout", {
 	bubbles: true,
 })
 /**
@@ -168,13 +152,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @param {HTMLElement} item The clicked item.
  * @public
  */
-@event<TreeItemClickEventDetail>("item-click", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("item-click", {
 	bubbles: true,
 	cancelable: true,
 })
@@ -187,13 +165,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @param {HTMLElement} item the deleted item.
  * @public
  */
-@event<TreeItemDeleteEventDetail>("item-delete", {
-	detail: {
-		/**
-		 * @public
-		 */
-		item: { type: HTMLElement },
-	},
+@event("item-delete", {
 	bubbles: true,
 })
 
@@ -202,10 +174,7 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @param {HTMLElement} item The focused item.
  * @private
  */
-@event<TreeItemFocusEventDetail>("item-focus", {
-	detail: {
-		item: { type: HTMLElement },
-	},
+@event("item-focus", {
 	bubbles: true,
 })
 
@@ -217,39 +186,28 @@ type WalkCallback = (item: TreeItemBase, level: number, index: number) => void;
  * @param {HTMLElement} targetItem The item triggering the event.
  * @public
  */
-@event<TreeSelectionChangeEventDetail>("selection-change", {
-	detail: {
-		/**
-		 * @public
-		 */
-		selectedItems: { type: Array },
-		/**
-		 * @public
-		 */
-		previouslySelectedItems: { type: Array },
-		/**
-		 * @public
-		 */
-		targetItem: { type: HTMLElement },
-	},
+@event("selection-change", {
 	bubbles: true,
 })
-@event<TreeMoveEventDetail>("move", {
-	detail: {
-		source: { type: Object },
-		destination: { type: Object },
-	},
+@event("move", {
 	bubbles: true,
 })
-@event<TreeMoveEventDetail>("move-over", {
-	detail: {
-		source: { type: Object },
-		destination: { type: Object },
-	},
+@event("move-over", {
 	bubbles: true,
 	cancelable: true,
 })
 class Tree extends UI5Element {
+	eventDetails!: {
+		"item-toggle": TreeItemToggleEventDetail,
+		"item-mouseover": TreeItemMouseoverEventDetail,
+		"item-mouseout": TreeItemMouseoutEventDetail,
+		"item-click": TreeItemClickEventDetail,
+		"item-delete": TreeItemDeleteEventDetail,
+		"item-focus": TreeItemFocusEventDetail,
+		"selection-change": TreeSelectionChangeEventDetail,
+		"move": TreeMoveEventDetail,
+		"move-over": TreeMoveEventDetail,
+	}
 	/**
 	 * Defines the selection mode of the component. Since the tree uses a `ui5-list` to display its structure,
 	 * the tree modes are exactly the same as the list modes, and are all applicable.
@@ -408,58 +366,23 @@ class Tree extends UI5Element {
 			return;
 		}
 
-		let placements = closestPosition.placements;
-
 		closestPosition.element = <HTMLElement>(<ShadowRoot>closestPosition.element.getRootNode()).host;
-
 		if (draggedElement.contains(closestPosition.element)) { return; }
-
 		if (closestPosition.element === draggedElement) {
-			placements = placements.filter(placement => placement !== MovePlacement.On);
+			closestPosition.placements = closestPosition.placements.filter(placement => placement !== MovePlacement.On);
 		}
 
-		const placementAccepted = placements.some(placement => {
-			const closestElement = closestPosition.element;
-			const beforeItemMovePrevented = !this.fireDecoratorEvent<TreeMoveEventDetail>("move-over", {
-				source: {
-					element: draggedElement,
-				},
-				destination: {
-					element: closestElement,
-					placement,
-				},
-			});
-
-			if (beforeItemMovePrevented) {
-				e.preventDefault();
-				this.dropIndicatorDOM!.targetReference = closestElement;
-				this.dropIndicatorDOM!.placement = placement;
-				return true;
-			}
-
-			return false;
-		});
-
-		if (!placementAccepted) {
-			this.dropIndicatorDOM!.targetReference = null;
-		}
+		const { targetReference, placement } = handleDragOver(e, this, closestPosition, closestPosition.element);
+		this.dropIndicatorDOM!.targetReference = targetReference;
+		this.dropIndicatorDOM!.placement = placement;
 	}
 
 	_ondrop(e: DragEvent) {
-		e.preventDefault();
-
-		const draggedElement = DragRegistry.getDraggedElement()!;
-		this.fireDecoratorEvent<TreeMoveEventDetail>("move", {
-			source: {
-				element: draggedElement,
-			},
-			destination: {
-				element: this.dropIndicatorDOM!.targetReference!,
-				placement: this.dropIndicatorDOM!.placement,
-			},
-		});
-		draggedElement.focus();
-		this.dropIndicatorDOM!.targetReference = null;
+		if (!this.dropIndicatorDOM?.targetReference || !this.dropIndicatorDOM?.placement) {
+			return;
+		}
+		handleDrop(e, this, this.dropIndicatorDOM.targetReference, this.dropIndicatorDOM.placement);
+		this.dropIndicatorDOM.targetReference = null;
 	}
 
 	_onListItemStepIn(e: CustomEvent<TreeItemBaseStepInEventDetail>) {
@@ -482,7 +405,7 @@ class Tree extends UI5Element {
 
 	_onListItemToggle(e: CustomEvent<TreeItemBaseToggleEventDetail>) {
 		const treeItem = e.detail.item;
-		const defaultPrevented = !this.fireDecoratorEvent<TreeItemToggleEventDetail>("item-toggle", { item: treeItem });
+		const defaultPrevented = !this.fireDecoratorEvent("item-toggle", { item: treeItem });
 		if (!defaultPrevented) {
 			treeItem.toggle();
 		}
@@ -491,26 +414,26 @@ class Tree extends UI5Element {
 	_onListItemClick(e: CustomEvent<ListItemClickEventDetail>) {
 		const treeItem = e.detail.item as TreeItemBase;
 
-		if (!this.fireDecoratorEvent<TreeItemClickEventDetail>("item-click", { item: treeItem })) {
+		if (!this.fireDecoratorEvent("item-click", { item: treeItem })) {
 			e.preventDefault();
 		}
 	}
 
 	_onListItemDelete(e: CustomEvent<ListItemDeleteEventDetail>) {
 		const treeItem = e.detail.item as TreeItemBase;
-		this.fireDecoratorEvent<TreeItemDeleteEventDetail>("item-delete", { item: treeItem });
+		this.fireDecoratorEvent("item-delete", { item: treeItem });
 	}
 
 	_onListItemFocus(e: CustomEvent<ListItemFocusEventDetail>) {
 		const treeItem = e.detail.item as TreeItemBase;
-		this.fireDecoratorEvent<TreeItemFocusEventDetail>("item-focus", { item: treeItem });
+		this.fireDecoratorEvent("item-focus", { item: treeItem });
 	}
 
 	_onListItemMouseOver(e: MouseEvent) {
 		const target = e.target;
 
 		if (this._isInstanceOfTreeItemBase(target)) {
-			this.fireDecoratorEvent<TreeItemMouseoverEventDetail>("item-mouseover", { item: target });
+			this.fireDecoratorEvent("item-mouseover", { item: target });
 		}
 	}
 
@@ -518,7 +441,7 @@ class Tree extends UI5Element {
 		const target = e.target;
 
 		if (this._isInstanceOfTreeItemBase(target)) {
-			this.fireDecoratorEvent<TreeItemMouseoutEventDetail>("item-mouseout", { item: target });
+			this.fireDecoratorEvent("item-mouseout", { item: target });
 		}
 	}
 
@@ -534,7 +457,7 @@ class Tree extends UI5Element {
 			item.selected = true;
 		});
 
-		this.fireDecoratorEvent<TreeSelectionChangeEventDetail>("selection-change", {
+		this.fireDecoratorEvent("selection-change", {
 			previouslySelectedItems,
 			selectedItems,
 			targetItem,